word-pattern-remplacement/requirements.txt
python-docx
pandas
python-dotenv
pyinstaller
openpyxl
libssl